[lib] Use the actions queue in the peer to peer context
Summary:
In D13563 we've introduced a regression described in https://linear.app/comm/issue/ENG-9470/mitigate-risks-of-effects-running-on-outdated-data#comment-48bb85bb which caused sending messages to sometimes stuck. This diff replaces the approach with a queue.

https://linear.app/comm/issue/ENG-9470/mitigate-risks-of-effects-running-on-outdated-data

Test Plan:
Sent a lot of messages in quick successions and haven't encountered any issues.
Tested leaving thick threads and the original issue still reproduces - looking for a solution.

Reviewers: kamil, ashoat

Reviewed By: kamil, ashoat

Differential Revision: https://phab.comm.dev/D13698