[lib][native][web] Add support for rendering array of messages in RobotextChatMessageInfoItem
Summary:
This diff makes it possible to render a robotext that corresponds to a list of messages.

We don't actually render a list of messages anywhere yet.

Depends on D13685

Test Plan: Tested in combination with the following diffs, where I construct `RobotextChatMessageInfoItem` with an array of `messageInfos` to represent combined messages ([ENG-9559](https://linear.app/comm/issue/ENG-9559/combine-some-kinds-of-robotext-when-they-appear-together))

Reviewers: tomek

Reviewed By: tomek

Differential Revision: https://phab.comm.dev/D13686