[lib] Add accountMissingStatus
Summary:
issue: [[https://linear.app/comm/issue/ENG-9630/add-accountmissingstatus-to-auxuserinfo | ENG-9630]]
`missingSince` will be set to the fist timestamp when the device list fails to be fetched. `lastChecked` will be updated with every next request that fails (the device list is not returned)

Test Plan: iTested with the next diffs. Tested by dispatching an action and setting those values

Reviewers: tomek, kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D13726