[lib] Reduce users missing from identity
Summary:
issue: [[https://linear.app/comm/issue/ENG-9631/update-lastchecked-and-missingsince-when-querying-identity | ENG-9631]]

Test Plan:
Tested by dispatching `setMissingDeviceListsActionType` and `setPeerDeviceListsActionType`.
Tested that when a user is missing a device list the `accountMissingStatus` is set. Checked that `lastChecked` is updated with next requests but `missingSince` stays the same.

Checked that if non-empty a device list is returned by identity, `accountMissingStatus` is removed.
Reviewers:
tomek, kamil
Subscribers: