[lib] Update usages of userHasDeviceList to acknowledge accountMissingStatus
Summary: issue: [[https://linear.app/comm/issue/ENG-9633/if-lastchecked-too-old-query-identity-again | ENG-9633]]

Test Plan:
Tested that when a user is missing a device list the `accountMissingStatus` is set. Checked that `lastChecked` is updated with next requests but `missingSince` stays the same.

checked that if non-empty a device list is returned by identity, `accountMissingStatus` is removed.

checked that further requests for the same user are stopped for 20s.

checked that if a user that has a recent `lastChecked` is picked in search (triggering `onSelectUserFromSearch` which calls `useUsersSupportThickThreads`) they are not queried for

Reviewers: tomek, kamil

Reviewed By: kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D13729