[native] Fix reacting to robotext when sidebar is disabled
Summary:
While working on the parent diffs, I noticed an issue where pressing on a robotext did nothing.

Turns out this is not a recent regression introduced by my changes, and is an issue that has been present since we refactored how the reaction options appear.

The issue is that we had a condition that exited if `visibleEntryIDs.length === 0`, but our reaction options aren't rendered via `visibleEntryIDs`.

This diff addresses it by adding `&& !canCreateReactionFromMessage` to the check.

Depends on D13708

Test Plan: Confirm that on native, when I press on a robotext in a sidebar, I'm able to see options to react

Reviewers: tomek, angelika, kamil

Subscribers: