[lib] Reduce removePeerUsersActionType in thread reducer
Summary:
issue: [[https://linear.app/comm/issue/ENG-9634/reduce-removepeerusersactiontype-in-thread-store | ENG-9634]]
In the thread reducer we want to use `generateOpsAndProcessThreadUpdates` which reduces delete account updates. In the future `removePeerUsersActionType` will be fully replaced by an update sent with `processDMOpsActionType` action. But this is tracked in [[https://linear.app/comm/issue/ENG-4905/support-flow-when-user-a-deletes-the-account | ENG-4905]]

Test Plan:
Tested with the next diff: tested that a user whos device list failed to be fetched for a long time gets deleted and they get removed from threads, users store and aux user store.

Reviewers:
tomek, kamil
Subscribers: