Fix payloadSizeExceeded calculated incorrectly
Summary:
`notificationSizeValidator` returns `true` if notification size is less than maximum notification size (that is, the size is correct). However we set `payloadSizeExceeded`
to `true` if `notificationSizeValidator` returns `true`. That's obviously wrong and we need to negate it.
I think the bug was introduced in the first place by `notificationSizeValidator` being not obvious name. I renamed it in another diff.

Test Plan:
1. Create a new chat
2. Verify that notification is verified correctly, that is `payloadSizeExceeded` is false.

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D13712