[keyserver] Make it so Ashoat creates chat with users on registration
Summary:
This diff mostly renames some variables. The only significant change is that we use my viewer to construct the thread with the new user, instead of using the user's viewer.

This is so that the robotext for the thread creation indicates that I created the thread, rather than indicating that the new user created the thread. The old behavior is a little bit confusing / misleading since the user never chose to create the thread.

Test Plan: [screenshot]

Reviewers: will, varun

Subscribers: