[lib] Show sidebars in top level if parent thread is in a separate tab
Summary:
This shows sidebars on top level if their parent threads have a differing notif setting (home/mute).

In the next diff, the sidebar items listed under a parent that exist as top level items in a different tab will be removed

Test Plan:
In a future diff, allowed the notif setting to be changed for sidebars. Tested the following

- Confirmed that moving the child thread had it appear in a separate tab as a top level item
- Confirmed that having a sidebar thread and its parent in the same tab did not result in a top level sidebar thread appearing
- Confirmed moving the parent into the same tab caused the child to no longer appear as top level and into a different tab resulted in a top level item

Reviewers: ashoat, varun

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D13719