[lib] Do not show sidebar underneath parent if in different tab
Summary:
We do not wish to show the sidebars underneath the parent if they already exist as top level items in a different tab.

Depends on D13719

Test Plan:
Tested by moving a child thread to a different tab and confirming that it disappeared underneath the parent thread.

Moving the parent thread to a separate tab from its child also resulted in the child disappearing from underneath.

Reviewers: ashoat, varun

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D13720