[keyserver] Delete Farcaster channel tag blobs when deleting a community
Summary:
In [ENG-9704](https://linear.app/comm/issue/ENG-9704/farcaster-channel-books-linked-to-deleted-community), we found that Farcaster channel tag blobs were not deleted when a community was deleted. This diff addresses that.

Depends on D13733

Test Plan: I created a community and tagged it with a Farcaster channel. I then deleted the community, and confirmed that the Farcaster channel tag blob was deleted in DynamoDB

Reviewers: tomek

Differential Revision: https://phab.comm.dev/D13734