[lib] Don't show Farcaster prompt for users after they unlink Farcaster
Summary: If the user unlinks their Farcaster identity, we should set it to `NO_FID_METADATA` so they don't get prompted again.

Test Plan: I looked through the code for places that call `useCurrentUserFID` to see where `undefined`/`null` is differentiated. I saw that `ConnectFarcasterAlertHandler` calls `shouldSkipConnectFarcasterAlert`, which differentiates the two. I confirmed that setting `NO_FID_METADATA` is better, so that we return `null` and so that `ConnectFarcasterAlertHandler` doesn't prompt the user to connect their Farcaster account in this case

Reviewers: will

Reviewed By: will

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D13735