[lib] Factor code out for updating FID in Redux
Summary:
In the next diff, we'll add some code to invalidate the `UserIdentityCache` for the current user when this happens. We want to do this whenever we update the FID locally, so it makes sense to factor all those locations out first.

Depends on D13735

Test Plan: Flow; this is a simple refactor

Reviewers: will

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D13736