[lib] Invalidate UserIdentityCache when changing FID
Summary: When we change the current user's FID, we should invalidate their entry in the `UserIdentityCache`. Otherwise, `FarcasterDataHandler` will continue to use the old value, and can end up wiping the FID if `handleCurrentUserFID` is run again before the app is killed (this can happen if the app is backgrounded/foregrounded).

This addresses [ENG-9704](https://linear.app/comm/issue/ENG-9704/farcaster-channel-books-linked-to-deleted-community).

Depends on D13736

Test Plan:
1. Open mobile app when Farcaster account not linked
2. Link Farcaster account
3. Background the app
4. Foreground the app
5. Confirm that the Farcaster account is not unlinked

Reviewers: will

Subscribers: