[web] update EditThreadAvatarMenu to add farcaster option
Summary: this lets a user set a farcaster thread avatar if the thread belongs to a community tagged with a farcaster channel

Depends on D13749

Test Plan: successfully set a farcaster avatar in community root thread and child threads. was able to unset the avatar as well.

menu icon looks the same as in the user avatar menu

Reviewers: ashoat

Subscribers: