[keyserver] create message and sidebar based on cast
Summary:
This creates a message based on a cast and converts it to a sidebar. If there is a parent cast included in the webhook event, we use the parent cast, if not, we use the tagger's cast.

Depends D13815

Example sidebar creation with commbot
{F3103594}

Test Plan: Triggered the creation of a new community with a webhook POST request and confirmed sidebar was correctly created in the community

Reviewers: ashoat, varun

Reviewed By: ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D13782