[keyserver] create invite link based on newly created sidebar
Summary:
This creates an invite link for the newly created sidebar.

Depends on D13775

Test Plan:
console logged the invitelink. On my iOS simulator dev build, plugged in the invite link to safari with `https://comm.app/invite/<value>` and confirmed I successfully joined the community and sidebar with a user originally not in either.

Reviewers:
varun,
ashoat

Subscribers: