[lib][keyserver][native] Stop sending passwords to authoritative keyserver on registration
Summary:
Context for why we want to revert this is in D12641. Linear task: [ENG-8698](https://linear.app/comm/issue/ENG-8698/revert-d12641)

This reverts commit 51a51ba64cc988a189e843d1165a8c77ecad578f.

It's not a perfect revert, since we need to preserve the field in `keyserverAuthRequestInputValidator` for old clients.

To keep the field in the validator, I had to split `KeyserverAuthRequest` into two types, one for `keyserver` and one for clients.

Test Plan: I ran through the registration flow in my local environment and made sure it still works

Reviewers: varun

Subscribers: