[keyserver] add function to grab comm userID for given fid if on keyserver
Summary:
This diff introduces the `getVerifiedUserIDForFID` function in a new file for farcaster utils on the keyserver.

This will be used later in the stack to determine whether or not the tagger or lead fids for a channel have comm accounts on the keyserver.

Test Plan:
flow check. This was tested later in the stack. It successfully returned a userID for an fid only if the userID existed on the keyserver.

Reviewers:
varun,
ashoat

Subscribers: