[keyserver] remove commbot after creating community
Summary: This removes commbot after it creates the community. There will always be another keysrver admin as it will either be the tagger, channel lead, or the keyserver admin.

Test Plan: Tested by auto-creating a community for the `dgn` channel and confirming that after commbot had created the community, the tagger userID was promoted to admin and commbot did not appear on members.

Reviewers: varun, ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D13813