[keyserver] check if viewer role is positive for updateThread
Summary:
Was running into an issue where because the viewer was constructed as a script viewer, it couldn't be added to the community thread with `joinThread` as `fetchViewerIsMember` always returns that script viewers are members. This diff
instead replaces the is a member check with a check for if the viewer role is positive, which will allow us to add script viewers to the thread.

Discussion at https://linear.app/comm/issue/ENG-8457/create-comm-community-for-a-farcaster-channel-if-it-doesnt-already

Test Plan: Was able to successfully add commbot and a test comm user to the community thread. Subsequent attempts to add did not result in duplicate robotext

Reviewers: ashoat, varun

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D13815