[lib] Merge generating notifs data with processing operations
Summary:
In the next diff we're modifying the notifications data so that it contains the threads necessary to generate notifs. For some specs, threads need to represent the state after processing the operation, which means that `notificationsCreationData` would need to call `processDMOperation`, use returned updates, and apply them to the threads. This has two problems: we would call the `processDMOperation` twice for each operation (once top-level, once from `notificationsCreationData`), and there would be a cycle between `notificationsCreationData` and `processDMOperation`, which calls `notificationsCreationData`. To solve these two issue we need to merge the logic into one method.

https://linear.app/comm/issue/ENG-9823/fix-sending-notifs-about-leaving-a-thick-thread

Depends on D13842

Test Plan: Flow. Did a sanity check - sent a text message to a thick thread and verified that a notif was delivered.

Reviewers: kamil, angelika

Reviewed By: kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D13843