[keyserver] add a addViewerAsGhost option to createThread
Summary:
This implements a `addViewerAsGhost` option to createThread. We use this in the context of farcaster bot to allow commbot to create threads without being added as a member.

Depends on D13811

Test Plan:
Tested later in the stack. Confirmed that community threads and sidebar threads created by commbot did not have commbot visible in the members list.

Reviewers:
ashoat,
varun

Subscribers: