[keyserver] Run num admin change checks in updateRole only if request is not by an admin
Summary:
Ran into an issue with running updateRole with a ghost commbot viewer to set the `communityAdminID` for `createTaggedFarcasterCommunity` causing a `invalid_parameters` as the changed admin count was equal to the current number of community admins.
This diff allows us to skip over this check if the request is to change members to admins

Depends on D13833

Test Plan:
Was able to successfully update the role for the prospective comm user admin to an actual admin as confirmed by the iOS simualtor's chat list without encountering `invalid_parameters`.

Reviewers:
ashoat,
varun

Subscribers: