[native_rust_library] implement method to retrieve latest backup info
Summary:
[ENG-6145](https://linear.app/comm/issue/ENG-6145/change-backup-restore-endpoint-to-expect-username-or-wallet-address).

This method will be used to het both `backupID` and `userID`. There is some code duplication but later in the stack other function like `retrieve_latest_siwe_backup_data` or `download_latest_backup_id` are deprecated in favour of `download_latest_backup_info` (I had to somehow split this into diffs to make it easier for a review). This function will be the main interface to get the backup latest info for both password and wallet users.

Depend on D13836

Test Plan: Tested later in the stack

Reviewers: varun, bartek