[native] unify username and wallet backup restore
Summary:
[ENG-6145](https://linear.app/comm/issue/ENG-6145/change-backup-restore-endpoint-to-expect-username-or-wallet-address).

This allows us to decrease code duplication (next diff) and use the same interface for wallet/username users.

Additionally, moving restore methods to `BackupMenu` as this is for now only for testing purposes and we'll need something different for actual backup.

Depends on D13840

Test Plan: Tested in D13849

Reviewers: varun, bartek

Reviewed By: varun, bartek

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D13844