[CommCoreModule][native_rust_library] update downloading User Keys to support user identifier
Summary:
[ENG-6145](https://linear.app/comm/issue/ENG-6145/change-backup-restore-endpoint-to-expect-username-or-wallet-address).

Alternative to this, is to pass `username` and get `BackupID` from Rust - in terms of network calls it should be the same. I implemented this because required a lot less changes because it was reusing existing code.

Alternatively, we can download only `backupID` and read `backupDataKey` and `backupLogDataKey` (it should exist on primary always) to avoid downloading User Keys - but curious for @bartek perspective as I don't fully remember what it was implemented that way,

1. Deprecate `download_latest_backup_id`
2. Update `retrieveBackupKeys` to have `backupID`

Depends on D13847

Test Plan:
This is the final diff in the stack so testing this end-to-end:
1. iOS and Android (I was using physical devices)
2. username and wallet user
3. testing restore similarly to [link](https://linear.app/comm/issue/ENG-9579/[biggest-risk]-test-existing-client-backup-code#comment-8001ee14)

TODO: test Secondary Auth Flow

Reviewers: varun, bartek

Subscribers: ashoat, tomek, bartek

Differential Revision: https://phab.comm.dev/D13849