[keyserver] add link back to original warpcast cast on farcaster bot sidebar source message
Summary:
Following discussion in https://linear.app/comm/issue/ENG-9838/add-a-link-back-to-warpcast-in-sidebar-source-message#comment-54f1ff96, we decided to give Comm users context to Comm taggings on Farcaster with a link back to the original cast.

This diff adds a link to "said" in the sidebar source message.

When one copies the cast hash directly on warpcast through the "Copy cast hash" option in a cast's upper right menu, it results in a link like so: `https://warpcast.com/ashoat.eth/0xccbf1cc1`, but it looks like using the full cast's hash in the link works just the same.

Test Plan:
Tested by running on my local keyserver and triggering a sidebar creation with a manual webhook event. Included a full cast hash and confirmed the resulting message viewed from my iOS simulator comm app had a linkified "said" word that when clicked, opened the cast on warpcast.

Reviewers:
ashoat,
varun

Subscribers: