[keyserver] change channel text from channel name to slash channel id
Summary:
Addresses https://linear.app/comm/issue/ENG-9847/use-slash-for-channel-instead-of-channel-name-in-sidebar-source.

Replace the "in channel <channel name>" text in the sidebar source message
with "in channel /<channel id>"

Test Plan: Tested with local keyserver and native app running on iOS simulator. Confirmed text appeared as `in channel /comm` in sidebar source message

Reviewers: ashoat, varun

Reviewed By: ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D13865