[keyserver] replace sidebar source message quotes with markdown quotes
Summary:
Addresses https://linear.app/comm/issue/ENG-9848/replace-quotes-with-markdown-quotes-in-sidebar-soucre-message and https://linear.app/comm/issue/ENG-9846/move-in-channel-channel-earlier-in-sidebar-source-message

In this diff, we replace the usage of quotations surrounding the cast text and replace it with markdown quotes by prepending `>` to each line of the original cast text.

The source message ordering is also rearranged with the channel text now accompanying the "author said" text which is then followed by the cast text.

Depends on D13865

Test Plan:
Tested by running local keyserver and triggering a sidebar creation with a webhook event. Confirmed that the sidebar source message appeared as expected.
{F3149092}

Reviewers: ashoat, varun

Reviewed By: ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D13866