[native] Consider BottomSheetHandle size when setting snapPoints
Summary: The height of `BottomSheetHandle` is always the same, and we should take it into account when setting `snapPoints`. More context [here](https://phab.comm.dev/D13125#387223).

Test Plan: In combination with the rest of the stack, I tested each individual use of our `BottomSheet` component to make sure that the bottom padding was reasonable. In most cases it was unchanged; in other cases, it was changed to be more consistent with the other `BottomSheet`s in our codebase. In all cases, there was enough padding for the "home pill" at the bottom of the screen

Reviewers: bartek

Subscribers: