[native] Measure whole KeyserverSelectionBottomSheet
Summary:
Instead of measuring just part of this, and adding up some hardcoded constants to the measured part, we can simply measure the whole thing.

In combination with the parent diff, this allows us to remove the hardcoded constants entirely, and rely purely on the measurement.

Depends on D13867

Test Plan: In combination with the rest of the stack, I tested each individual use of our `BottomSheet` component to make sure that the bottom padding was reasonable. In most cases it was unchanged; in other cases, it was changed to be more consistent with the other `BottomSheet`s in our codebase. In all cases, there was enough padding for the "home pill" at the bottom of the screen

Reviewers: bartek

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D13868