[native] Fix incorrectly set verticalBounds in UserProfileAvatar
Summary:
`verticalBounds` of `FullScreenViewModal` are using to establish a border that will "cut off" part of the image as it animates back.

In the chat view, we use to prevent the animating image from appearing over the tab bar and the header.

In his initial implementation of `UserProfileAvatar`, Ginsu seems to have set `verticalBounds` based on an uninformed guess as to its function.

We don't have any requirements for the profile bottom sheet to have a border that cuts off the image as it animates back, since the bottom sheet appears above (over) navigational elements such as the tab bar and the header.

As such, we can effectively disable `verticalBounds` by setting it to be the full screen.

This fixes a visual issue where part of the user avatar gets cut off when animating back.

Depends on D13868

Test Plan: I confirmed the visual glitch was gone

Reviewers: bartek

Reviewed By: bartek

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D13869