[native] Get rid of userProfileBottomPadding
Summary:
This padding was hardcoded by Ginsu to add additional space at the bottom of the `BottomSheet`. It is obviated by my work to consider `BottomSheetHandle` when setting `snapPoints`.

This diff reduces the padding somewhat, to be consistent with the rest of our `BottomSheet`s.

Depends on D13869

Test Plan: In combination with the rest of the stack, I tested each individual use of our `BottomSheet` component to make sure that the bottom padding was reasonable. In most cases it was unchanged; in other cases, it was changed to be more consistent with the other `BottomSheet`s in our codebase. In all cases, there was enough padding for the "home pill" at the bottom of the screen

Reviewers: bartek

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D13870