[native] Get rid of bottomSheetPaddingTop in ConnectFarcasterBottomSheet
Summary:
This padding was hardcoded by Ginsu to add additional space at the bottom of the `BottomSheet`. It is obviated by my work to consider `BottomSheetHandle` when setting `snapPoints`.

The padding was extra large because Ginsu measured the height of the button incorrectly. Once I adjusted the height of button, the padding is approximately the same as what it was before.

Depends on D13870

Test Plan: In combination with the rest of the stack, I tested each individual use of our `BottomSheet` component to make sure that the bottom padding was reasonable. In most cases it was unchanged; in other cases, it was changed to be more consistent with the other `BottomSheet`s in our codebase. In all cases, there was enough padding for the "home pill" at the bottom of the screen

Reviewers: bartek

Subscribers: